Change to using <ul> to display list
