+++ src/App.css
@@ -37,12 +37,3 @@
   top: 50px;
   left: 50%;
 }
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
